% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_gene_examination.R
\name{gene_exam}
\alias{gene_exam}
\title{Exploratory huva experiment}
\usage{
gene_exam(
  huva_dataset = huva_default_data,
  gene = gene_name,
  bins = 30,
  alpha = 1
)
}
\arguments{
\item{huva_dataset}{huva_dataset class object.}

\item{gene}{gene name used for the analysis.}

\item{bins}{graphical option, see ggplot2 for details.}

\item{alpha}{graphical option, see ggplot2 for details.}
}
\value{
huva_gene_examination class object including a graphical overview of the distribution of gene expression of a selected GOI.
}
\description{
This funtion looks into the expression of a gene of interest (GOI) in the specified dataset or across other user-provided datasets.
}
\details{
Examine the profile of expression of a selected gene and investigate the relationship with annotation parameters of
metadata
}
\examples{
library(huva)
library(huva.db)

gene_overview <- gene_exam(huva_dataset = huva.db, gene = "MYD88")

}
\seealso{
run_huva_experiment
}
