% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_misc.R
\name{generate_huva_dataset}
\alias{generate_huva_dataset}
\title{Preparation of a new huva_dataset with independent data}
\usage{
generate_huva_dataset(
  dataset_name,
  data,
  annotation,
  metadata = NULL,
  combine = F
)
}
\arguments{
\item{dataset_name}{charachter name for the new huva dataset generated}

\item{data}{dataset data}

\item{annotation}{dataset annotation}

\item{metadata}{if NULL no metadata will be added}

\item{combine}{LOGICAL, if TRUE the new dataset will be combined with the huva default dataset}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Preparation of a new huva_dataset with independent data
}
\examples{
# This will generate a combined dataset with new data and the data provided in the huva package

2+2

}
