% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_anno.plot_exam}
\alias{get_anno.plot_exam}
\title{Plot of annotation table parameters for a selected gene of interest}
\usage{
get_anno.plot_exam(huva_expression, study = "ALL", dataset = "ALL")
}
\arguments{
\item{huva_expression}{huva_expression class object}

\item{study}{charachter or charachter vector defining the studies to be used in the analysis (default = "ALL")}

\item{dataset}{charachter or charachter vector defining the datasets to be used in the analysis (default = "ALL")}
}
\value{
data.frame or list of the annotation data of a selected gene of interest across the available datasets
}
\description{
Plot of annotation table parameters for a selected gene of interest
}
\examples{

2+2

}
