% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_anno.stat_huva}
\alias{get_anno.stat_huva}
\title{get_anno.stats_huva}
\usage{
get_anno.stat_huva(huva_exp, study, dataset = NULL, easytoread = T)
}
\arguments{
\item{huva_exp}{huva experiment class object}

\item{study}{Chaarachter vector defininf the names of the study to be visualised}

\item{dataset}{if NULL (default) all datasets will be used}

\item{easytoread}{if TRUE (default) transforms all output in data.frame ojects}
}
\value{
return statistics list on annotation table
}
\description{
get_anno.stats_huva
}
\examples{

2+2

}
