% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_DE_huva}
\alias{get_DE_huva}
\title{get_DE_huva}
\usage{
get_DE_huva(
  huva_exp,
  study,
  dataset = NULL,
  pval = 0.05,
  logFC = 0,
  cluster_col = T,
  PC = c("PC1", "PC2")
)
}
\arguments{
\item{huva_exp}{huva experiment class object}

\item{study}{character vector defining the names of the study to be visualized}

\item{dataset}{if NULL (default) all datasets will be used}

\item{pval}{p-value cut-off used for the definition of differential expression}

\item{logFC}{log2 fold change cut-off for the definition of differential expression}

\item{cluster_col}{TRUE defaul}

\item{PC}{character vetor defining the PC to be plotted (c("PC1", "PC2"), default)}
}
\value{
table of deferentially expressed genes in the comparison between the binned samples
}
\description{
get_DE_huva
}
\examples{

2+2

}
