% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_expr_exam}
\alias{get_expr_exam}
\title{Expression table for a selected gene of interest}
\usage{
get_expr_exam(huva_expression, study, dataset)
}
\arguments{
\item{huva_expression}{huva_expression class object}

\item{study}{charachter or charachter vector defining the studies to be used in the analysis (default = "ALL")}

\item{dataset}{charachter or charachter vector defining the datasets to be used in the analysis (default = "ALL")}
}
\value{
data.frame or list of the expression of a selected gene of interest across the available datasets
}
\description{
Expression table for a selected gene of interest
}
\examples{

2+2

}
