% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_gsea_huva}
\alias{get_gsea_huva}
\title{get_gsea_huva}
\usage{
get_gsea_huva(huva_exp, study, dataset = NULL, int_plot = T)
}
\arguments{
\item{huva_exp}{huva experiment class object.}

\item{study}{character vector defining the name of the studies to be visualised.}

\item{dataset}{if NULL (default) all datasets will be used.}

\item{int_plot}{default is TRUE and provides an interactive GSEA Volcano plot.}
}
\value{
return the GSEA table as a data frame and a Volcano plot (NES in x-axis, -log10pval in y-axis).
}
\description{
Gene set enrichment analysis (GSEA) is performed on user-defined huva experiment ranked lists. By default, the huva experiment works on hallmark gene sets.
}
\details{
get_gsea_huva
}
\examples{
library(huva)
library(huva.db)

binned_dataset <- run_huva_experiment(data = huva.db,
                                     gene = "MYD88",
                                     quantiles = 0.10,
                                     gs_list = hallmarks_V7.2,
                                     summ = TRUE,
                                     datasets_list = NULL,
                                     adjust.method = "none")

gsea_huva <- get_gsea_huva(huva_exp = binned_dataset,
                           study = "FG500")

gsea_huva$plot_FG500_whole_blood

gsea_huva$int_plot_FG500_whole_blood

}
\seealso{
run_huva_experiment
}
