% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_meta.plot_exam}
\alias{get_meta.plot_exam}
\title{get_meta.plot_exam}
\usage{
get_meta.plot_exam(huva_expression, study = "ALL", dataset = "ALL")
}
\arguments{
\item{huva_expression}{huva_expression class object.}

\item{study}{character vector defining the studies to be used in the analysis (default = "ALL").}

\item{dataset}{character vector defining the datasets to be used in the analysis (default = "ALL").}
}
\value{
The function returns a plot list based on the annotation table.
}
\description{
The get_meta.plot_exam function will quickly provide a list of metadata-based graphical objects in
             a huva examination experiment.
}
\details{
Metadata table parameters for a selected gene of interest
}
\examples{
library(huva)
library(huva.db)

gene_overview <- gene_exam(huva_dataset = huva.db, gene = "MYD88")

meta.plot_exam <- get_meta.plot_exam(huva_expression = gene_overview,
                                     study = "FG500",
                                     dataset = "FG500_whole_blood_cellcount")

meta.plot_exam$`Monocytes (CD14+)`

}
\seealso{
gene_exam, run_huva_experiment
}
