% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_meta.stat_huva}
\alias{get_meta.stat_huva}
\title{get_meta.stats_huva}
\usage{
get_meta.stat_huva(huva_exp, study, dataset = NULL)
}
\arguments{
\item{huva_exp}{huva experiment class object.}

\item{study}{character vector defining the name of the studies to be visualised.}

\item{dataset}{if NULL (default) all datasets will be used.}
}
\value{
returns a list of statistics performed on the annotation table.
}
\description{
Within the two huva experiment groups, statistical differences in metadata parameters can be explored with the function get_meta.stat.huva.
}
\details{
get_meta.stats_huva
}
\examples{
library(huva)
library(huva.db)

binned_dataset <- run_huva_experiment(data = huva.db,
                                     gene = "MYD88",
                                     quantiles = 0.10,
                                     gs_list = hallmarks_V7.2,
                                     summ = TRUE,
                                     datasets_list = NULL,
                                     adjust.method = "none")

meta.stat <- get_meta.stat_huva(huva_exp = binned_dataset,
                                study = "FG500",
                                dataset = "FG500_whole_blood_cellcount")

}
\seealso{
run_huva_experiment
}
