% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_meta_exam}
\alias{get_meta_exam}
\title{get_meta_exam}
\usage{
get_meta_exam(huva_expression, study = "ALL", dataset = "ALL")
}
\arguments{
\item{huva_expression}{huva_expression class object.}

\item{study}{character vector defining the studies to be used in the analysis (default = "ALL").}

\item{dataset}{character vector defining the datasets to be used in the analysis (default = "ALL").}
}
\value{
data.frame or list of the annotation data of a selected gene of interest across the available datasets.
}
\description{
The function provides metadata table parameters for a selected gene of interest used in a huva
             gene examination experiment.
}
\details{
Metadata table parameters for a selected gene of interest
}
\examples{
library(huva)
library(huva.db)

gene_overview <- gene_exam(huva_dataset = huva.db,
                           gene = "MYD88")

meta.table_exam <- get_meta_exam(huva_expression = gene_overview,
                                 study = "FG500")

}
\seealso{
gene_exam, run_huva_experiment
}
