% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_meta_huva}
\alias{get_meta_huva}
\title{get_meta_huva}
\usage{
get_meta_huva(huva_exp, study, dataset = NULL)
}
\arguments{
\item{huva_exp}{huva experiment class object.}

\item{study}{character vector defining the name of the studies to be visualised.}

\item{dataset}{if NULL (default) all datasets will be used.}
}
\value{
annotation table of binned samples.
}
\description{
The function get_meta_huva returns the metadata table of the specified huva_experiment groups.
}
\details{
get_meta_huva
}
\examples{
library(huva)
library(huva.db)

binned_dataset <- run_huva_experiment(data = huva.db,
                                     gene = "MYD88",
                                     quantiles = 0.10,
                                     gs_list = hallmarks_V7.2,
                                     summ = TRUE,
                                     datasets_list = NULL,
                                     adjust.method = "none")

meta_huva <- get_meta_huva(huva_exp = binned_dataset,
                           study = "FG500")

}
\seealso{
run_huva_experiment
}
