% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_plotting.R
\name{plot_binned_gene}
\alias{plot_binned_gene}
\title{Plot selected genes from an huva_experiment class object}
\usage{
plot_binned_gene(goi, huva_experiment)
}
\arguments{
\item{goi}{Name of the gene of interest}

\item{huva_experiment}{huva_experiment class object}
}
\value{
list of ggplot objects, each plot recapitulates gene expression of selected gene/s in the present datasets
}
\description{
Plot selected genes from an huva_experiment class object
}
\examples{

2+2

}
