% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_experiment.R
\name{run_huva_phenotype}
\alias{run_huva_phenotype}
\title{Calculate huva experiment on the variance of a seleted metadata paramiter}
\usage{
run_huva_phenotype(
  data,
  phenotype,
  study,
  metadata_table,
  quantiles,
  gs_list,
  summ = T,
  adjust.method = "none"
)
}
\arguments{
\item{data}{huva_dataset class object.}

\item{phenotype}{gene name used for the analysis}

\item{study}{defines the study to be used in the analysis}

\item{metadata_table}{name of the metadata table in wich the selected paramiter can be found}

\item{quantiles}{definition of the quantile of segregation of the samples, quantiles are always simmetrical
between high and low groups (e.g. \code{quantile 0.1} will use the 10th and 90th percentiles)}

\item{gs_list}{class list object defining gene sets to be included in the analysis (to generate this file see
the documentation of fgse)}

\item{summ}{default TRUE, define if the summary of the huva experiment will be calculated}

\item{adjust.method}{p value adjustment method used to corret the DE genes analysis}
}
\value{
huva_experiment based on the segregaation of the phenotype
}
\description{
Calculate huva experiment on the variance of a seleted metadata paramiter
}
\examples{

2+2

}
