% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_gene_examination.R
\name{gene_exam}
\alias{gene_exam}
\title{Examin the profile of expression of a selected gene and ivestigate the relation with annotation paramiters of
metadata}
\usage{
gene_exam(
  huva_dataset = huva_default_data,
  gene = gene_name,
  bins = 30,
  alpha = 1
)
}
\arguments{
\item{huva_dataset}{huva_dataset class object.}

\item{gene}{gene name used for the analysis}

\item{bins}{graphica option, see ggplot2}

\item{alpha}{see ggplot2}
}
\value{
huva_gene_examination class object, it includes and overview of the distribution of gene expression for a selected GOI.
}
\description{
Examin the profile of expression of a selected gene and ivestigate the relation with annotation paramiters of
metadata
}
\examples{

2+2

}
