% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_misc.R
\name{generate_huva_dataset}
\alias{generate_huva_dataset}
\title{generate_huva_dataset}
\usage{
generate_huva_dataset(
  dataset_name,
  data,
  annotation,
  metadata = NULL,
  combine = F
)
}
\arguments{
\item{dataset_name}{character name for the new huva dataset.}

\item{data}{dataset input data.}

\item{annotation}{dataset annotation table.}

\item{metadata}{if NULL no metadata will be added.}

\item{combine}{LOGICAL, if TRUE the new dataset will be combined with the huva default dataset.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function provides a new dataset including expression data, sample annotation
             and possible metadata to run a huva experiment with novel data.
}
\details{
Preparation of a new huva_dataset with user-defined data
}
\examples{
library(huva)
library(huva.db)

data_FG <- huva.db$FG500$data
anno_FG <- huva.db$FG500$anno
meta_FG <- huva.db$FG500$metadata

new_huva.db <- generate_huva_dataset(dataset_name = "newdata",
                                     data = "data_FG",
                                     annotation = "anno_FG",
                                     metadata = "meta_FG",
                                     combine = FALSE)


}
\seealso{
run_huva_experiment
}
