% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_anno_huva}
\alias{get_anno_huva}
\title{get_anno_huva}
\usage{
get_anno_huva(huva_exp, study, dataset = NULL)
}
\arguments{
\item{huva_exp}{huva experiment class object}

\item{study}{Chaarachter vector defininf the names of the study to be visualised}

\item{dataset}{if NULL (default) all datasets will be used}
}
\value{
annotation table of the binned samples
}
\description{
get_anno_huva
}
\examples{

2+2

}
