% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_expr.plot_exam}
\alias{get_expr.plot_exam}
\title{Visualization of the expression histogram for a selected gene of interest}
\usage{
get_expr.plot_exam(
  huva_expression,
  study = "ALL",
  dataset = "ALL",
  bins = 30,
  alpha = 1
)
}
\arguments{
\item{huva_expression}{huva_expression class object}

\item{study}{charachter or charachter vector defining the studies to be used in the analysis (default = "ALL")}

\item{dataset}{charachter or charachter vector defining the datasets to be used in the analysis (default = "ALL")}

\item{bins}{see ggplot2 documentation (default = 30)}

\item{alpha}{see ggplot2 documentation (default = 1)}
}
\value{
ggplot class object, histogram for the expression of a selected gene of interest across selected datasets
}
\description{
Visualization of the expression histogram for a selected gene of interest
}
\examples{

2+2

}
