% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_expr_huva}
\alias{get_expr_huva}
\title{get_expr_huva}
\usage{
get_expr_huva(huva_exp, study, dataset = NULL)
}
\arguments{
\item{huva_exp}{huva experiment class object.}

\item{study}{character vector defining the name of the studies to be visualised.}

\item{dataset}{if NULL (default) all datasets will be used.}
}
\value{
expression table of binned samples as data frame object.
}
\description{
The function get_expr_huva provides the expression table resulting from reference huva experiment data.frame.
}
\details{
get_expr_huva
}
\examples{
library(huva)
library(huva.db)

binned_dataset <- run_huva_experiment(data = huva.db,
                                     gene = "MYD88",
                                     quantiles = 0.10,
                                     gs_list = hallmarks_V7.2,
                                     summ = TRUE,
                                     datasets_list = NULL,
                                     adjust.method = "none")

expr_huva <- get_expr_huva(huva_exp = binned_dataset,
                           study = "FG500",
                           dataset = "FG500_whole_blood")

}
\seealso{
run_huva_experiment
}
