% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_gsea_huva}
\alias{get_gsea_huva}
\title{get_gsea_huva}
\usage{
get_gsea_huva(huva_exp, study, dataset = NULL, int_plot = T)
}
\arguments{
\item{huva_exp}{huva experiment class object}

\item{study}{Chaarachter vector defininf the names of the study to be visualised}

\item{dataset}{if NULL (default) all datasets will be used}

\item{int_plot}{default TRUE, decide if the interactive GSEA plot will be calculated}
}
\value{
return GSEA table and plot
}
\description{
get_gsea_huva
}
\examples{

2+2

}
