% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_meta.plot_huva}
\alias{get_meta.plot_huva}
\title{get_meta.plot_huva}
\usage{
get_meta.plot_huva(huva_exp, study, dataset = NULL)
}
\arguments{
\item{huva_exp}{huva experiment class object.}

\item{study}{character vector defining the name of the studies to be visualised.}

\item{dataset}{if NULL (default) all datasets will be used.}
}
\value{
The function returns a plot list on the metadata table.
}
\description{
The graphical representation of the distribution of individuals across the two huva groups can be visualiuzed in correlation to metadata parameters with the get_anno.plot_huva function.
}
\details{
get_meta.plot_huva
}
\examples{
library(huva)
library(huva.db)

binned_dataset <- run_huva_experiment(data = huva.db,
                                     gene = "MYD88",
                                     quantiles = 0.10,
                                     gs_list = hallmarks_V7.2,
                                     summ = TRUE,
                                     datasets_list = NULL,
                                     adjust.method = "none")

meta.plot <- get_meta.plot_huva(huva_exp = binned_dataset,
                                study = "FG500")

meta.plot$FG500_whole_blood_cellcount$Granulocytes

}
\seealso{
run_huva_experiment
}
