% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_examination.R
\name{get_meta.stat_exam}
\alias{get_meta.stat_exam}
\title{Metadata table parameters for a selected gene of interest}
\usage{
get_meta.stat_exam(
  huva_expression,
  study = "ALL",
  dataset = "ALL",
  param = "ALL"
)
}
\arguments{
\item{huva_expression}{huva_expression class object}

\item{study}{charachter or charachter vector defining the studies to be used in the analysis (default = "ALL")}

\item{dataset}{charachter or charachter vector defining the datasets to be used in the analysis (default = "ALL")}

\item{param}{colName of the sample table to be used for the statistical analyisis}
}
\value{
data.frame or list of the annotation data of a selected gene of interest across the available datasets
}
\description{
Metadata table parameters for a selected gene of interest
}
\examples{

2+2

}
