% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_explore_huva.R
\name{get_rank_huva}
\alias{get_rank_huva}
\title{get_rank_huva}
\usage{
get_rank_huva(huva_exp, study, dataset = NULL, n_top_genes = 20)
}
\arguments{
\item{huva_exp}{huva experiment class object}

\item{study}{Chaarachter vector defininf the names of the study to be visualised}

\item{dataset}{if NULL (default) all datasets will be used}

\item{n_top_genes}{number of top down or up regulated genes to be displayed}
}
\value{
ranking of the genes
}
\description{
get_rank_huva
}
\examples{

2+2

}
