% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_misc.R
\name{plot_HM}
\alias{plot_HM}
\title{Plot heatmap object}
\usage{
plot_HM(HM)
}
\arguments{
\item{HM}{object containing output from the pheatmap function. Created by assigning output of
pheatmap function to an object (i.e. pheatmap_object <- pheatmap(matrix))}
}
\value{
A heatmap created by pheatmap
}
\description{
Plot heatmap object
}
\examples{
2+2
}
\keyword{plot_pheatmap}
