% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_experiment.R
\name{run_huva_experiment}
\alias{run_huva_experiment}
\title{Calculate huva experiment on the expression variance of a selected gene of interest}
\usage{
run_huva_experiment(
  data = datasets,
  gene,
  quantiles,
  gs_list,
  summ = T,
  datasets_list = NULL,
  adjust.method = "none"
)
}
\arguments{
\item{data}{huva_dataset class object.}

\item{gene}{gene name used for the analysis}

\item{quantiles}{definition of the quantile of segregation of the samples, quantiles are always simmetrical
between high and low groups (e.g. \code{quantile 0.1} will use the 10th and 90th percentiles)}

\item{gs_list}{class list object defining gene sets to be included in the analysis (to generate this file see
the documentation of fgse)}

\item{summ}{bolean. define if the summary of the huva experiment will be calculated}

\item{datasets_list}{characher vector used to filter the dataset in the data objet for the analysis}

\item{adjust.method}{p value adjustment method used to corret the DE genes analysis}
}
\value{
huva_experiment
}
\description{
Calculate huva experiment on the expression variance of a selected gene of interest
}
\examples{

2+2

}
