% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huva_experiment.R
\name{summary_huva_experiment}
\alias{summary_huva_experiment}
\title{Summarise huva experiment in list or string}
\usage{
summary_huva_experiment(
  huva_experiment,
  include_gene = T,
  include_gsea = T,
  include_metadata = F,
  gene_list,
  metadata_list = NULL,
  one_line_result = F
)
}
\arguments{
\item{huva_experiment}{huva_experiment class object calculate with the run_huva experiment function.}

\item{include_gene}{LOGIC. If TRUE genes fold change will be included in the summary}

\item{include_gsea}{LOGIC. If TRUE NES and p value from GSEA will be included in the analysis}

\item{include_metadata}{LOGIC. If TRUE metadata fold change in the two groups will be included in the summary}

\item{gene_list}{Filter for genes to be included in the summary (charachter vector)}

\item{metadata_list}{Filter for metadata tables to be included in the anaysis}

\item{one_line_result}{LOGIC. If TRUE the summary will be transformed in a single string}
}
\value{
summary of huva_experiment
}
\description{
Summarise huva experiment in list or string
}
\examples{

2+2

}
